Add admin rights + check for duplicate passwords

+++ PPl3/Areas/Admin/Controllers/HomeAdminController.cs
@@ -1,19 +1,28 @@
-﻿using PPl3.Models;
-using System;
+﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Net.Mail;
 using System.Web;
 using System.Web.Mvc;
-using static System.Collections.Specialized.BitVector32;
+using System.Web.Routing;
 using System.Web.Security;
-using System.IO;
-using System.Net.Mail;
+using Newtonsoft.Json;
+using PPl3.Models;
+using PPl3.App_Start;
+using System.Globalization;
+using System.Text.RegularExpressions;
+using System.Configuration;
+using PPl3.Models.Payments;
+using System.Data.Entity;
+using Microsoft.AspNet.SignalR;
+using PPl3.Hubs;
+using System.Text;
+using static System.Data.Entity.Infrastructure.Design.Executor;
+using System.Security.Policy;
 using System.Net;
-using System.Web.Management;
-using System.Data;
-using WebGrease;
-using System.Web.UI;
-
+using System.Security.Cryptography;
+using System.Web.UI.WebControls;
 namespace PPl3.Areas.Admin.Controllers
 {
     public class HomeAdminController : Controller
@@ -21,39 +30,48 @@ namespace PPl3.Areas.Admin.Controllers
         // GET: Admin/HomeAdmin
         public ActionResult Index()
         {
-            if (TempData["CheckComfirm"] != null)
+            if (Session["user"] != null)
             {
-                ViewBag.CheckComfirm = 1;
-            }
-            PPL3Entities db = new PPL3Entities();
-            string result1 = "";
-            string result2 = "";
-            int i = 1;
-            while (i <= 12)
-            {
-                int count1 = 0;
-                int count2 = 0;
-                foreach (var item in db.users)
+                if (TempData["CheckComfirm"] != null)
                 {
-                    if (item.user_type == 3)
+                    ViewBag.CheckComfirm = 1;
+                }
+                PPL3Entities db = new PPL3Entities();
+                string result1 = "";
+                string result2 = "";
+                int i = 1;
+                while (i <= 12)
+                {
+                    int count1 = 0;
+                    int count2 = 0;
+                    foreach (var item in db.users)
                     {
-                        if(item.day_become_host.Value.Month == i) count1++; 
+                        if (item.user_type == 3)
+                        {
+                            if (item.day_become_host.Value.Month == i) count1++;
+                        }
+
                     }
-                    
+                    foreach (var item in db.properties)
+                    {
+                        if (item.Date_Post.Value.Month == i && item.p_status == 1) count2++;
+                    }
+                    result1 = result1 + count1.ToString() + ",";
+                    result2 = result2 + count2.ToString() + ",";
+                    i++;
                 }
-                foreach(var item in db.properties)
-                {
-                    if (item.Date_Post.Value.Month == i && item.p_status == 1) count2++;
-                }
-                result1 = result1 + count1.ToString() + ",";
-                result2 = result2 + count2.ToString() + ",";
-                i++;
+                ViewBag.data_hosts = result1.ToString().Substring(0, result1.Length - 1);
+                ViewBag.data_hotels = result2.ToString().Substring(0, result2.Length - 1);
+                return View();
             }
-            ViewBag.data_hosts = result1.ToString().Substring(0, result1.Length - 1);
-            ViewBag.data_hotels = result2.ToString().Substring(0, result2.Length - 1);
-            return View();
+            else
+            {
+                return RedirectToAction("login", "HomeUser", new { area = "user" });
+            }
+            
         }
         [HttpPost]
+        [AdminAuhorize(idChucNang = 17)]
         public JsonResult DeleteUser(int id)
         {
             PPL3Entities db = new PPL3Entities();
@@ -69,6 +87,7 @@ namespace PPl3.Areas.Admin.Controllers
             }
             try
             {
+                db.Messages.RemoveRange(db.Messages.Where(item => item.SenderID == id || item.ReceiverID == id).ToList());
                 db.host_reviews.RemoveRange(db.host_reviews.Where(item => item.hostid == id || item.review_by_user == id).ToList());
                 db.Friendships.RemoveRange(db.Friendships.Where(item => item.UserId2 == id || item.UserId1 == id).ToList());
                 db.transactions.RemoveRange(db.transactions.Where(item => item.receiver_id == id || item.payer_id == id).ToList());
@@ -88,6 +107,7 @@ namespace PPl3.Areas.Admin.Controllers
         }
 
         [HttpPost]
+        [AdminAuhorize(idChucNang = 17)]
         public JsonResult DeleteHotel(int id)
         {
             PPL3Entities db = new PPL3Entities();
@@ -102,7 +122,8 @@ namespace PPl3.Areas.Admin.Controllers
                 phanHoi = e.ToString();
             }
             return Json(phanHoi, JsonRequestBehavior.AllowGet);
-        }       
+        }
+        [AdminAuhorize(idChucNang = 15)]
         public ActionResult confirm_host(int user_id)
         {
             PPL3Entities db = new PPL3Entities();
@@ -126,6 +147,7 @@ namespace PPl3.Areas.Admin.Controllers
             return RedirectToAction("index");
         }
         [HttpPost]
+        [AdminAuhorize(idChucNang = 15)]
         public JsonResult host_denied(int user_id)
         {
             PPL3Entities db = new PPL3Entities();
@@ -145,7 +167,7 @@ namespace PPl3.Areas.Admin.Controllers
             return Json("true", JsonRequestBehavior.AllowGet);
         }
 
-
+        [AdminAuhorize(idChucNang = 16)]
         public ActionResult confirm_hotel(int hotel_id , int user_id)
         {
             PPL3Entities db = new PPL3Entities();
@@ -170,8 +192,8 @@ namespace PPl3.Areas.Admin.Controllers
             return RedirectToAction("index");
         }
 
-        
 
+        [AdminAuhorize(idChucNang = 16)]
         public JsonResult hotel_denied(int hotel_id, int user_id)
         {
             PPL3Entities db = new PPL3Entities();
@@ -188,18 +210,19 @@ namespace PPl3.Areas.Admin.Controllers
             return Json("true", JsonRequestBehavior.AllowGet);
         }
 
-       
+        [AdminAuhorize(idChucNang = 15)]
         public ActionResult ComfirmAll(int[] id, int check)
 
         {
             PPL3Entities db = new PPL3Entities();
-            if (check == 1)
+            if (check == 1 && id.Length > 0)
             {
                 List<browser_becomes_host> list_host = new List<browser_becomes_host>();
                 for (int i = 0; i <= id.Length - 1; ++i)
                 {
                     int find_id = id[i];
                     browser_becomes_host find_host = db.browser_becomes_host.FirstOrDefault(item => item.user_id == find_id);
+                    if(find_host != null)
                     list_host.Add(find_host);
                 }
                 foreach (var item in list_host)
@@ -225,16 +248,17 @@ namespace PPl3.Areas.Admin.Controllers
                 }
                 db.SaveChanges();
                 TempData["CheckComfirm"] = 1;
-                return RedirectToAction("index");
+              
             }
-            else
+            else if (check == 2 && id.Length > 0)
             {
-                
+
                 List<Browse_hotel_listings> list_hotel = new List<Browse_hotel_listings>();
                 for (int i = 0; i <= id.Length - 1; ++i)
                 {
                     int find_id = id[i];
                     Browse_hotel_listings find_hotel = db.Browse_hotel_listings.FirstOrDefault(item => item.property_id == find_id);
+                    if(find_hotel != null)
                     list_hotel.Add(find_hotel);
                 }
                 foreach (var item in list_hotel)
@@ -259,23 +283,24 @@ namespace PPl3.Areas.Admin.Controllers
                 }
                 db.SaveChanges();
                 TempData["CheckComfirm"] = 1;
-                return RedirectToAction("index");
+               
             }
+            return RedirectToAction("index");
         }
-
+        [AdminAuhorize(idChucNang = 15)]
         public ActionResult DeniedAll(int[] id , int check)
         {
             PPL3Entities db = new PPL3Entities();
-            if (check == 1)
+            if (check == 1 && id.Length > 0)
             {
                 List<browser_becomes_host> list_host = new List<browser_becomes_host>();
                 for (int i = 0; i <= id.Length - 1; ++i)
                 {
                     int find_id = id[i];
                     browser_becomes_host find_host = db.browser_becomes_host.FirstOrDefault(item => item.user_id == find_id);
-                    list_host.Add(find_host);
+                    if(find_host != null) list_host.Add(find_host);
                 }
-               
+
                 foreach (var item in list_host)
                 {
                     db.browser_becomes_host.Remove(item);
@@ -284,7 +309,7 @@ namespace PPl3.Areas.Admin.Controllers
                 TempData["CheckComfirm"] = 1;
                 return RedirectToAction("index");
             }
-            else
+            else if (check == 2 && id.Length > 0)
             {
 
                 List<Browse_hotel_listings> list_hotel = new List<Browse_hotel_listings>();
@@ -292,9 +317,10 @@ namespace PPl3.Areas.Admin.Controllers
                 {
                     int find_id = id[i];
                     Browse_hotel_listings find_hotel = db.Browse_hotel_listings.FirstOrDefault(item => item.property_id == find_id);
+                    if(find_hotel != null)
                     list_hotel.Add(find_hotel);
                 }
-               
+
                 foreach (var item in list_hotel)
                 {
                     db.Browse_hotel_listings.Remove(item);
@@ -303,6 +329,7 @@ namespace PPl3.Areas.Admin.Controllers
                 TempData["CheckComfirm"] = 1;
                 return RedirectToAction("index");
             }
+            else return RedirectToAction("index");
         } 
         public ActionResult LogOut()
         {
